frontend/requirements.txt
bottle >= 0.12.7
Paste >= 3.0.8
google_cloud_datastore >= 1.7.3
